scan: improved scan_sampleMasterMix, PrescaleValue fix

+++ firmware_17xx/synth/scan.h
@@ -23,7 +23,7 @@ typedef void (*scan_event_callback_t)(int8_t source); // source: keypad (kb0..kb
 uint16_t scan_getPotValue(int8_t pot);
 void scan_resetPotLocking(void);
 void scan_setMode(int8_t isSmpMasterMixMode);
-void scan_sampleMasterMix(uint16_t sampleCount, uint8_t * buffer);
+void scan_sampleMasterMix(uint16_t sampleCount, uint16_t * buffer);
 void scan_setScanEventCallback(scan_event_callback_t callback);
 
 int scan_potTo16bits(int x);
